+++ compliance/flexera/iam/iam_explicit_user_roles/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v3.2
+
+- Updated policy metadata to make it more clear what Flexera service the policy is for
+
 ## v3.1
 
 - Updated description to account for new file path in Github repository